requirements.txt
django
djangorestframework
markdown
django-registration
django-cors-middleware
freezegun
Pillow
python-magic
django-extensions
